feat: fix up types and Ora
