+++ packages/cli-doctor/src/tools/downloadAndUnzip.ts
@@ -1,5 +1,5 @@
 import {fetchToTemp} from '@react-native-community/cli-tools';
-import {Loader} from '@react-native-community/cli-types';
+import {Loader} from '../types';
 import {unzip} from './unzip';
 import {deleteFile} from './deleteFile';
