+++ packages/cli-doctor/src/tools/healthchecks/__tests__/androidSDK.test.ts
@@ -5,8 +5,8 @@ import {cleanup, writeFiles} from '../../../../../../jest/helpers';
 import androidSDK from '../androidSDK';
 import getEnvironmentInfo from '../../envinfo';
 import * as downloadAndUnzip from '../../downloadAndUnzip';
-import {EnvironmentInfo} from '@react-native-community/cli-types';
-import {NoopLoader} from '../../../tools/loader';
+import {EnvironmentInfo} from '../../../types';
+import {NoopLoader} from '@react-native-community/cli-tools';
 import * as common from '../common';
 import * as androidWinHelpers from '../../windows/androidWinHelpers';
 import * as environmentVariables from '../../windows/environmentVariables';
