+++ packages/cli-doctor/src/tools/healthchecks/common.ts
@@ -2,8 +2,8 @@ import chalk from 'chalk';
 import readline from 'readline';
 import wcwidth from 'wcwidth';
 import stripAnsi from 'strip-ansi';
-import {Loader} from '@react-native-community/cli-types';
 import {logger} from '@react-native-community/cli-tools';
+import {Loader} from '../../types';
 
 // Space is necessary to keep correct ordering on screen
 const logMessage = (message?: string) => {
