+++ packages/cli-doctor/src/tools/healthchecks/watchman.ts
@@ -1,7 +1,7 @@
 import versionRanges from '../versionRanges';
 import {doesSoftwareNeedToBeFixed} from '../checkInstallation';
 import {install} from '../install';
-import {HealthCheckInterface} from '@react-native-community/cli-types';
+import {HealthCheckInterface} from '../../types';
 
 const label = 'Watchman';
