+++ packages/cli-types/src/index.ts
@@ -12,11 +12,7 @@ import {
   AndroidDependencyParams,
 } from './android';
 
-import type {Ora} from 'ora';
-
 export type Prompt = any;
-
-export type Loader = Ora;
 
 export type CommandFunction<Args = Object> = (
   argv: Array<string>,
@@ -124,99 +120,6 @@ export type ProjectConfig = {
   [key: string]: any;
 };
 
-export type NotFound = 'Not Found';
-type AvailableInformation = {
-  version: string;
-  path: string;
-};
-
-type Information = AvailableInformation | NotFound;
-
-export type EnvironmentInfo = {
-  System: {
-    OS: string;
-    CPU: string;
-    Memory: string;
-    Shell: AvailableInformation;
-  };
-  Binaries: {
-    Node: AvailableInformation;
-    Yarn: AvailableInformation;
-    npm: AvailableInformation;
-    Watchman: AvailableInformation;
-  };
-  SDKs: {
-    'iOS SDK': {
-      Platforms: string[];
-    };
-    'Android SDK':
-      | {
-          'API Levels': string[] | NotFound;
-          'Build Tools': string[] | NotFound;
-          'System Images': string[] | NotFound;
-          'Android NDK': string | NotFound;
-        }
-      | NotFound;
-  };
-  IDEs: {
-    'Android Studio': AvailableInformation | NotFound;
-    Emacs: AvailableInformation;
-    Nano: AvailableInformation;
-    VSCode: AvailableInformation;
-    Vim: AvailableInformation;
-    Xcode: AvailableInformation;
-  };
-  Languages: {
-    Java: Information;
-  };
-};
-
-export type HealthCheckCategory = {
-  label: string;
-  healthchecks: HealthCheckInterface[];
-};
-
-export type Healthchecks = {
-  common: HealthCheckCategory;
-  android: HealthCheckCategory;
-  ios?: HealthCheckCategory;
-};
-
-export type RunAutomaticFix = (args: {
-  loader: Loader;
-  logManualInstallation: ({
-    healthcheck,
-    url,
-    command,
-    message,
-  }: {
-    healthcheck?: string;
-    url?: string;
-    command?: string;
-    message?: string;
-  }) => void;
-  environmentInfo: EnvironmentInfo;
-}) => Promise<void> | void;
-
-export type HealthCheckInterface = {
-  label: string;
-  visible?: boolean | void;
-  isRequired?: boolean;
-  description?: string;
-  getDiagnostics: (
-    environmentInfo: EnvironmentInfo,
-  ) => Promise<{
-    version?: string;
-    versions?: [string];
-    versionRange?: string;
-    needsToBeFixed: boolean | string;
-  }>;
-  win32AutomaticFix?: RunAutomaticFix;
-  darwinAutomaticFix?: RunAutomaticFix;
-  linuxAutomaticFix?: RunAutomaticFix;
-  runAutomaticFix: RunAutomaticFix;
-};
-
 /**
  * @property root - Root where the configuration has been resolved from
  * @property reactNativePath - Path to React Native source
@@ -249,7 +152,7 @@ export interface Config extends IOSNativeModulesConfig {
     [name: string]: PlatformConfig<any, any, any, any>;
   };
   commands: Command[];
-  healthChecks: HealthCheckCategory[];
+  healthChecks: [];
 }
 
 /**
@@ -275,7 +178,7 @@ export type UserDependencyConfig = {
   // An array of extra platforms to load
   platforms: Config['platforms'];
   // Additional health checks
-  healthChecks: HealthCheckCategory[];
+  healthChecks: [];
 };
 
 export {
